+++ crosscompute/scripts/tools/see.py
@@ -6,28 +6,28 @@ from ...routines import (
     get_crosscompute_token)
 
 
-class SeeProjectScript(LoggingScript):
+class SeeToolScript(LoggingScript):
 
     def configure(self, argument_subparser):
         super().configure(argument_subparser)
         argument_subparser.add_argument(
-            'project-id', metavar='PROJECT-ID', nargs='?')
+            'tool-id', metavar='TOOL-ID', nargs='?')
 
     def run(self, args, argv):
         super().run(args, argv)
         host = get_crosscompute_host()
         token = get_crosscompute_token()
-        project_id = getattr(args, 'project-id')
-        d = run(host, token, project_id)
+        tool_id = getattr(args, 'tool-id')
+        d = run(host, token, tool_id)
         return d
 
 
-def run(host, token, project_id=None):
-    url = host + '/projects'
-    if not project_id:
+def run(host, token, tool_id=None):
+    url = host + '/tools'
+    if not tool_id:
         url += '.json'
     else:
-        url += f'/{project_id}.json'
+        url += f'/{tool_id}.json'
     headers = {'Authorization': 'Bearer ' + token}
     response = requests.get(url, headers=headers)
     return response.json()